LineBot/requirements.txt
astroid==2.1.0
beautifulsoup4==4.6.3
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.0.2
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
pylint==2.2.2
requests==2.21.0
six==1.12.0
typed-ast==1.1.1
urllib3==1.24.1
virtualenv==16.1.0
Werkzeug==0.14.1
wrapt==1.10.11
flask-cors
gunicorn
pytz
line-bot-sdk
flask_cors
configparser